Main controller
For some reason the changes I made to main controller, were changed back to before I made changes. Rewrote the files.
